organized all imports according according to new package structure in jackson-syntax-highlight
