+++ pom.xml
@@ -48,7 +48,7 @@
 		<maven.compiler.target>1.8</maven.compiler.target>
 
 		<logstash-logback-encoder.version>6.6</logstash-logback-encoder.version>
-		<jackson-syntax-highlight.version>1.0.5</jackson-syntax-highlight.version>
+		<jackson-syntax-highlight.version>1.0.6-SNAPSHOT</jackson-syntax-highlight.version>
 		<commons-text.version>1.9</commons-text.version>
 		<jackson.version>2.12.1</jackson.version>
 		<logback.version>1.2.3</logback.version>
