+++ src/main/java/com/github/skjolber/decorators/factory/LogLevelSyntaxHighlighterFactory.java
@@ -2,9 +2,9 @@ package com.github.skjolber.decorators.factory;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.github.skjolber.decorators.syntaxhighlight.LogLevelSyntaxHighlighter;
-import com.github.skjolber.jackson.jsh.AnsiSyntaxHighlight;
-import com.github.skjolber.jackson.jsh.DefaultSyntaxHighlighter;
-import com.github.skjolber.jackson.jsh.SyntaxHighlighter;
+import com.github.skjolber.jackson.jsh.highlighter.AnsiSyntaxHighlight;
+import com.github.skjolber.jackson.jsh.highlighter.DefaultSyntaxHighlighter;
+import com.github.skjolber.jackson.jsh.highlighter.SyntaxHighlighter;
 
 public class LogLevelSyntaxHighlighterFactory extends ConfigurableSyntaxHighlighterFactory {
